backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
automata-lib==9.1.2
bcrypt==4.2.0
cached_method==0.1.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.120.2
frozendict==2.4.6
git-filter-repo==2.47.0
h11==0.16.0
idna==3.11
networkx==3.5
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pymongo==4.15.3
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.1
typing-inspection==0.4.2
typing_extensions==4.14.1
uvicorn==0.38.0
